backend/requirements.txt
fastapi
uvicorn
motor  # Cliente async para MongoDB
pydantic
python-dotenv
paho-mqtt
websockets
wsproto